refactor: rename filesystem.py to xdg.py
