+++ src/ghsnitch/logger.py
@@ -1,6 +1,6 @@
 import logging
 
-from .filesystem import LOG_DIR
+from .xdg import LOG_DIR
 
 _LOG_FILE = LOG_DIR / "run.log"
 _FMT = "%(asctime)s %(levelname)-8s %(name)s: %(message)s"
